@ggerganov 2023-10-27
@slaren Have you noticed this as well? Do you think there is any reason not to switch for F32 compute?