@slaren 2023-10-27
I tested this again on a 3090 Ti, and for me master is faster:

  Device 0: NVIDIA GeForce RTX 3090 Ti, compute capability 8.6
| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 | pp 512     | 4672.50 ± 290.50 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 | tg 128     |     54.62 ± 0.18 |

build: c8d6a1f3 (1431) (master)


  Device 0: NVIDIA GeForce RTX 3090 Ti, compute capability 8.6
| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 | pp 512     |  3285.46 ± 11.15 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 | tg 128     |     55.63 ± 0.19 |

build: 3b9ea655 (1432) (PR)