@KerfuffleV2 2023-10-27
Might be a dumb question, but would these changes affect quantized models at all?