@ggerganov 2023-10-27
@slaren Indeed, on RTX 3090 I also don't observe benefit from 32F mode. So it's not an universal thing

@KerfuffleV2 Yes, because for some of the operations we dequantize to F16 and use cuBLAS. There are some numbers in my post above for quantized models (`Q4_K`)