@ggerganov 2023-10-27
I restored the `CUDA_COMPUTE_16F` in `ggml_cuda_op_mul_mat_cublas()` (used for 2D tensors) and kept `CUDA_COMPUTE_32F` in `ggml_cuda_mul_mat_mat_batched_cublas()` (used for 3D tensors KV cache). Now the PP performance is maintained for 3090 and maybe slightly better for the others and the slight TG improvement is also maintained:

  Device 0: NVIDIA GeForce RTX 3090, compute capability 8.6
| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| llama 13B mostly Q4_K - Medium |   7.33 GiB |    13.02 B | CUDA       |  99 | pp 512     |  2303.96 ± 25.72 |
| llama 13B mostly Q4_K - Medium |   7.33 GiB |    13.02 B | CUDA       |  99 | tg 128     |     69.43 ± 0.24 |

build: 0f2498f (1433)

```bash
LLAMA_CUBLAS=1 make -j batched-bench && ./batched-bench ./models/llama-2-13b.Q4_K_M.gguf 4096 1 99 1 512,3200 128,800 1

### master

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.248 |  2060.39 |    1.985 |    64.48 |    2.234 |   286.52 |
|   512 |    800 |    1 |   1312 |    0.240 |  2131.52 |   12.681 |    63.08 |   12.922 |   101.54 |
|  3200 |    128 |    1 |   3328 |    1.946 |  1644.78 |    2.416 |    52.98 |    4.362 |   763.01 |
|  3200 |    800 |    1 |   4000 |    1.929 |  1658.93 |   15.530 |    51.51 |   17.459 |   229.11 |

### PR

main: n_kv_max = 4096, is_pp_shared = 1, n_gpu_layers = 99, mmq = 1

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|   512 |    128 |    1 |    640 |    0.243 |  2110.64 |    1.898 |    67.43 |    2.141 |   298.96 |
|   512 |    800 |    1 |   1312 |    0.233 |  2197.75 |   12.323 |    64.92 |   12.556 |   104.49 |
|  3200 |    128 |    1 |   3328 |    1.893 |  1690.68 |    2.452 |    52.21 |    4.344 |   766.03 |
|  3200 |    800 |    1 |   4000 |    1.872 |  1709.50 |   15.783 |    50.69 |   17.655 |   226.57 |
```

  Device 0: NVIDIA RTX A6000, compute capability 8.6
| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 | pp 512     |  4182.22 ± 93.96 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 | tg 128     |     45.23 ± 0.12 |

build: 0f2498f (1433)