@slaren 2023-10-27
The performance is the same now for me too, maybe a little bit better than master, but within the margin of error.

  Device 0: NVIDIA GeForce RTX 3090 Ti, compute capability 8.6
| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 | pp 512     |  4851.38 ± 70.18 |
| llama 7B mostly F16            |  12.55 GiB |     6.74 B | CUDA       |  99 | tg 128     |     55.83 ± 0.12 |

build: 0f2498f2 (1433)