@cebtenzzre 2023-10-28
Prompt processing results on my Tesla P40:
| GPU      | Model     | Test | t/s master | t/s PR | Speedup |
|----------|-----------|------|------------:|--------:|---------:|
| P40 | 7b q4_0   | pp512   |       338.89 |   934.11 |     2.76 |
| P40 | 13b q4_k_s  | pp512   |       204.26 |   472.83 |   2.31   |