@julianullrich99 2023-10-30
I saw this and was excited for better performance on my cheapo P40+1080ti setup. (Both Tesla Cards that have 1:64 in fp16 vs fp32, so should be waaayyy faster, i guess?)
However, I only saw a 50% improvement (20t/s -> 30t/s) on 13B mostly Q5_K_M
And even less with only 4.5t/s -> 5.5t/s on 70b Q3_K_M (the most vram i can fit in both cards (36gb).

What is the explanation for this bad improvement?