@julianullrich99 2023-10-30
I saw this and was excited for better performance on my cheapo P40+1080ti setup. (Both Tesla Cards that have 1:64 in fp16 vs fp32, so should be waaayyy faster, i guess?)
However, I only saw the following improvements:

| GPUs       | model                        | test  | t/s (master) | t/s (PR) | speedup |
|------------|------------------------------|-------|--------------|----------|---------|
| P40+1080ti | llama 13b mostly Q5_K medium | tg128 | 20.3         | 32.0     | 1.6     |
| P40+1080ti | llama 70b mostly Q3_K medium | tg128 | 4.3          | 5.7      | 1.3     |

What is the explanation for this poor improvement?

(q3 70b is the most I can fit in VRAM)