@cebtenzzre 2023-11-05
Updated performance figures (with -DLLAMA_CUDA_FORCE_MMQ=ON):
| GPU      | Model     | Test | t/s master | t/s PR | Speedup |
|----------|-----------|------|------------:|--------:|---------:|
| P40 | 7b q4_0   | pp512   |    889.54    |   934.11 |   1.05   |
| P40 | 13b q4_k_s  | pp512   |    455.17    |   472.83 |   1.04  |

It would still be cool if we could make this a command-line flag like mmq used to be.